pyproject.toml
[project]
name = "personal-agent"
version = "0.1.0"
description = "Personal AI Agent with WeChat integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "zhipuai>=2.0.0",
    "dashscope>=1.14.0",
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "playwright>=1.40.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "apscheduler>=3.10.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "rich>=13.0.0",
    "itchat>=1.3.0",
    "aiofiles>=23.0.0",
    "tenacity>=8.2.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "aiohttp>=3.9.0",
    "pycryptodome>=3.19.0",
    "zhdate>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
personal-agent = "personal_agent.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/personal_agent"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
